Refactor expression language tests
