+++ packages/expression_language/test/expressions_test.dart
@@ -1,7 +1,8 @@
 import 'dart:async';
 import 'package:gherkin/gherkin.dart';
 import 'package:glob/glob.dart';
-import 'supporting_files/steps/when_then_expression_num.dart';
+
+import 'supporting_files/index.dart';
 
 void main() async {
   final config = TestConfiguration()
@@ -18,6 +19,7 @@ void main() async {
     }
     ..stepDefinitions = [
       GivenFormElementIsProvided(),
+      GivenCustomExpressionIsProvided(),
       WhenExpressionIsEvaluated(),
       ThenIntExpressionResult(),
       ThenStringExpressionResult(),
