+++ packages/expression_language/test/features/custom_expression.feature
@@ -1,0 +1,7 @@
+Feature: Custom Expression
+  Tests Cusotm expression
+
+  Scenario: Custom Expression Test
+    Given custom expression is provided
+    When expression ""ab" + testExpression("cd","ef")" is evaluated
+    Then string expression result is "abcdef"
