+++ packages/expression_language/test/supporting_files/mocks/constant_expression_provider.dart
@@ -1,0 +1,12 @@
+import 'package:expression_language/expression_language.dart';
+
+class ConstantExpressionProvider<T> extends ExpressionProvider<T> {
+  final T value;
+
+  ConstantExpressionProvider(this.value);
+
+  @override
+  Expression<T> getExpression() {
+    return ConstantExpression(value);
+  }
+}
