+++ packages/expression_language/test/supporting_files/mocks/test_expression.dart
@@ -1,0 +1,26 @@
+import 'package:expression_language/expression_language.dart';
+
+class TestExpression extends Expression<String> {
+  final Expression<String> left;
+  final Expression<String> right;
+
+  TestExpression(this.left, this.right);
+
+  @override
+  Expression<String> clone(Map<String, ExpressionProviderElement> elementMap) {
+    return TestExpression(left, right);
+  }
+
+  @override
+  String evaluate() {
+    return left.evaluate() + right.evaluate();
+  }
+
+  @override
+  List<Expression> getChildren() {
+    return [
+      left,
+      right,
+    ];
+  }
+}
