+++ packages/expression_language/test/supporting_files/steps/given_form_element_is_provided.dart
@@ -1,0 +1,17 @@
+import 'package:gherkin/gherkin.dart';
+
+import '../index.dart';
+
+class GivenFormElementIsProvided extends GivenWithWorld<ExpressionWorld> {
+  @override
+  Future<void> executeStep() async {
+    world.buildGrammarParser(
+      {
+        'testElement': TestElement(),
+      },
+    );
+  }
+
+  @override
+  RegExp get pattern => RegExp(r'form element is provided');
+}
