+++ packages/expression_language/test/supporting_files/steps/then_bool_expression_result.dart
@@ -1,0 +1,13 @@
+import 'package:gherkin/gherkin.dart';
+
+import '../index.dart';
+
+class ThenBoolExpressionResult extends Then1WithWorld<String, ExpressionWorld> {
+  @override
+  Future<void> executeStep(String result) async {
+    expectMatch(world.result, result == 'true');
+  }
+
+  @override
+  RegExp get pattern => RegExp(r'bool expression result is {string}');
+}
