+++ packages/expression_language/test/supporting_files/steps/then_decimal_expression_result.dart
@@ -1,0 +1,15 @@
+import 'package:expression_language/expression_language.dart';
+import 'package:gherkin/gherkin.dart';
+
+import '../index.dart';
+
+class ThenDecimalExpressionResult
+    extends Then1WithWorld<String, ExpressionWorld> {
+  @override
+  Future<void> executeStep(String result) async {
+    expectMatch(world.result, Decimal.parse(result));
+  }
+
+  @override
+  RegExp get pattern => RegExp(r'decimal expression result is {string}');
+}
