+++ packages/expression_language/test/supporting_files/steps/then_string_expression_result.dart
@@ -1,0 +1,14 @@
+import 'package:gherkin/gherkin.dart';
+
+import '../index.dart';
+
+class ThenStringExpressionResult
+    extends Then1WithWorld<String, ExpressionWorld> {
+  @override
+  Future<void> executeStep(String result) async {
+    expectMatch(world.result, result);
+  }
+
+  @override
+  RegExp get pattern => RegExp(r'string expression result is {string}');
+}
